move Metamodel to spoon.metamodel and merge it with SpoonMetamodel
